pyproject.toml
[project]
name = "openai-agent-sdk-step-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.8.0",
    "openai-agents>=0.3.0",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
]
